Fix cloud deploy: pandas scalar comparison bug, limit price fetching to top 50
- Fix ValueError in mean_reversion_model: vol_avg was a pandas Series,
  not a scalar — cast to float before comparison
- Limit price history fetching to top 50 stocks by Composite score
  instead of entire universe (was timing out on Streamlit Cloud)
- Remove eager returns panel building from MODEL tab to prevent
  additional timeout on page load

+++ src/model/signals.py
@@ -436,8 +436,8 @@ def mean_reversion_model(
 
                     # Volume spike on down-move (capitulation indicator)
                     if "Volume" in ph.columns and len(ph) > 21:
-                        vol_recent = ph["Volume"].tail(5).mean()
-                        vol_avg = ph["Volume"].tail(63).mean()
+                        vol_recent = float(ph["Volume"].tail(5).mean())
+                        vol_avg = float(ph["Volume"].tail(63).mean())
                         is_down = ret_1m < 0
                         spike = (vol_recent / vol_avg - 1) if vol_avg > 0 else 0
                         vol_spike.append(spike if is_down else 0)
